api/urls/requirements.txt
aiohttp==3.6.2
aiohttp-jwt==0.5.0
aioredis==1.3.1
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
hiredis==1.0.1
idna==2.8
motor==2.1.0
multidict==4.7.2
pybase62==0.4.3
PyJWT==1.7.1
pymongo==3.10.0
PyYAML==5.2
trafaret==2.0.2
yarl==1.4.2
